+++ CHANGELOG.md
@@ -4,6 +4,9 @@ mdr1986x-small-lib changelog
 ## [Unreleased]
 ### Fixed
 - Using of incorrect function names in queue module
+### Changed
+- Type of data in queue module to uint8_t except of return values, because
+functions returns -1 when queue has no data.
 
 ## [0.2] - 2019-07-09
 ### Added